Move `SequenceGenerator` to `api.py`
