+++ outlines/generate/generator.py
@@ -19,65 +19,6 @@ class GenerationState:
     logits: torch.Tensor
 
 
-class SequenceGenerator:
-    def __init__(self, fsm, model, sampler, device):
-        self.generate_token = token_generator(model, sampler)
-        self.fsm = fsm
-        self.tokenizer = model.tokenizer
-        self.device = device
-
-    def format_sequence(self, sequence):
-        return sequence
-
-    def __call__(
-        self,
-        prompt,
-        kv_cache: Optional[torch.tensor] = None,
-        rng: Optional[torch.Generator] = None,
-    ) -> Union[str, List[str]]:
-        sequence_generator = self.stream(prompt, kv_cache, rng)
-        tokens = [token for token in sequence_generator]
-        sequences = [
-            self.format_sequence("".join(sequence)) for sequence in list(zip(*tokens))
-        ]
-        return sequences if len(sequences) > 1 else sequences[0]
-
-    def stream(
-        self,
-        prompt: str,
-        kv_cache: Optional[torch.tensor] = None,
-        rng: Optional[torch.Generator] = None,
-    ) -> Iterator[Union[List[str], str]]:
-        if rng is None:
-            rng = torch.Generator(device=self.device)
-            rng.seed()
-
-        init_state = init_generator_state(self.tokenizer, self.device, prompt, kv_cache)
-
-        token_ids = init_state[1]
-        num_sequences = token_ids.shape[0]
-
-        init_fsm_states = [FSMState(0) for _ in range(num_sequences)]
-
-        states = sequence_generator(
-            self.generate_token, self.fsm, init_state, init_fsm_states, rng
-        )
-
-        def token_generator() -> Iterator[Union[List[str], str]]:
-            while True:
-                try:
-                    sequence = next(states)
-                except StopIteration:
-                    return
-
-                next_token_ids = sequence.token_ids[:, -1]
-                next_tokens = self.tokenizer.decode(next_token_ids)
-
-                yield next_tokens
-
-        return token_generator()
-
-
 def init_generator_state(
     tokenizer: "Tokenizer",
     device: str,
